% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utilities.R
\name{assert_dataframe_has_column}
\alias{assert_dataframe_has_column}
\title{Assert Data Frame Has Column}
\usage{
assert_dataframe_has_column(df, param_name)
}
\arguments{
\item{df}{A data frame object to be validated.}

\item{param_name}{Character. The name of the column to check in the data frame.}
}
\value{
This function does not return a value. It stops execution with an error message if the
validations fail or if the column is not present.
}
\description{
Validates that the provided data frame contains the specified column. It uses
pre-existing checks to validate the data frame and the parameter name, then verifies
that the given column exists in the data frame.
}
\examples{
\dontrun{
  df <- data.frame(a = 1:5, b = 6:10)
  assert_dataframe_has_column(df, "a")
}

}
