% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utilities.R
\name{assert_is_valid_data_frame}
\alias{assert_is_valid_data_frame}
\title{Assert Valid Data Frame}
\usage{
assert_is_valid_data_frame(df)
}
\arguments{
\item{df}{A data frame object to be validated.}
}
\value{
This function does not return a value. It stops execution with an error message if any
of the validations fail.
}
\description{
Validates that the provided object is a valid data frame. It checks that the data frame is
not \code{NULL}, that it is of class \code{data.frame}, and that it has at least one column and one row.
}
\examples{
\dontrun{
df <- data.frame(x = 1:10)
assert_is_valid_data_frame(df)
}

}
