% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utilities.R
\name{assert_is_valid_parameter}
\alias{assert_is_valid_parameter}
\title{Assert Is Valid Parameter}
\usage{
assert_is_valid_parameter(param_name)
}
\arguments{
\item{param_name}{Character. The name of the parameter to validate.}
}
\value{
This function does not return a value. It stops execution with an error message if the validations fail.
}
\description{
Validates that the provided parameter name is a single character string and is not NA.
}
\examples{
\dontrun{
  assert_is_valid_parameter("age")
}

}
