% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utilities.R
\name{filter_complete}
\alias{filter_complete}
\title{Filter Complete}
\usage{
filter_complete(df, param_name)
}
\arguments{
\item{df}{A data frame object to filter.}

\item{param_name}{Character. The name of the column to check for non-\code{NA} values.}
}
\value{
A \code{data.frame} containing only rows with non-\code{NA} values in the specified column.
}
\description{
Filters the data frame by removing rows where the specified column contains \code{NA} values.
}
\examples{
\dontrun{
  df <- data.frame(a = c(1, NA, 3), b = c(4, 5, NA))
  complete_df <- filter_complete(df, "a")
}

}
