% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utilities.R
\name{stopifnot_with_message}
\alias{stopifnot_with_message}
\title{Assert Condition With Message}
\usage{
stopifnot_with_message(condition, message, ...)
}
\arguments{
\item{condition}{Logical. The condition to be evaluated.}

\item{message}{Character. The base message to be used if the condition is FALSE.}

\item{...}{Additional message parts to be concatenated for the error message.}
}
\value{
This function does not return a value. It stops execution with an error message if
the condition is FALSE.
}
\description{
Checks a condition and stops execution with a custom error message if the condition is not met.
The error message is formed by concatenating the additional arguments.
}
\examples{
\dontrun{
  x <- 5
  stopifnot_with_message(x > 10, "x must be greater than 10")
}

}
